+++ client/src/components/PlanDetails.js
@@ -1,6 +1,7 @@
 import React, { useContext } from 'react';
 
 import Table from 'react-bootstrap/Table';
+import ListGroup from 'react-bootstrap/ListGroup';
 
 import { CurrentPlanContext } from '../contexts/CurrentPlanContext';
 import { CustomisationContext } from '../contexts/CustomisationContext';
@@ -11,13 +12,13 @@ function PlanDetails() {
     const { currentPlan } = useContext(CurrentPlanContext);
     const { customOptions } = useContext(CustomisationContext);
 
-    return (
+    return (<>
         <Table responsive striped hover size="sm" variant={customOptions.displaySetting === "lightMode" ? "light" : "dark"}>
             <thead>
             <tr>
                 <th scope="col">#</th>
                 <th scope="col">Index/Waitlist/Vacancy</th>
-                <th scope="col">Course Code and Name</th>
+                <th scope="col">Course Code/Name</th>
                 <th scope="col">AUs</th>
                 <th scope="col">Exam Timing</th>
             </tr>
@@ -41,6 +42,16 @@ function PlanDetails() {
                 })}
             </tbody>
         </Table>
+        {currentPlan.length !== 0 ?
+            <ListGroup horizontal className="legend">
+                <ListGroup.Item className="flex-fill" variant="secondary">Legend:</ListGroup.Item>
+                <ListGroup.Item className="flex-fill">* Available as UEs</ListGroup.Item>
+                <ListGroup.Item className="flex-fill">~ Available as BDEs</ListGroup.Item>
+                <ListGroup.Item className="flex-fill">^ Self-paced Course</ListGroup.Item>
+                <ListGroup.Item className="flex-fill"># Available as GERPEs</ListGroup.Item>
+            </ListGroup> : null
+        }
+    </>
     )
 }