+++ client/src/components/overviews/IndexEditOverview.js
@@ -59,7 +59,7 @@ function IndexEditOverview() {
           <tr>
             <th scope="col">#</th>
             <th scope="col">Index/Waitlist/Vacancy</th>
-            <th scope="col">Course Code and Name</th>
+            <th scope="col">Course Code/Name</th>
             <th scope="col">AUs</th>
             <th scope="col">Exam Timing</th>
           </tr>